reworked stuff

+++ Server/pxe/create_iso.sh
@@ -8,9 +8,7 @@ git clone https://github.com/CCC-Jugendgruppe/Laptop-Management.git
 
 isopath="./Laptop-Management/iso"
 
-if [ ! -f "$isopath/debian12.iso" ]; then
-    wget https://cdimage.debian.org/debian-cd/current/amd64/iso-dvd/debian-12.1.0-amd64-DVD-1.iso -O "$isopath/debian12.iso"
-fi
+[ ! -f "$isopath/debian12.iso" ] && wget https://cdimage.debian.org/debian-cd/current/amd64/iso-dvd/debian-12.1.0-amd64-DVD-1.iso -O "$isopath/debian12.iso"
 
 cd "$isopath"
